pyproject.toml
[tool.poetry]
name = "crypto-impact-bot"
version = "0.1.0"
description = ""
authors = ["v-argunov <argunovvlad5@gmail.com>"]
readme = "README.md"
packages = [{include = "crypto_impact_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
pytelegrambotapi = "^4.8.0"
python-dotenv = "^0.21.0"
sqlalchemy = "^1.4.44"
pytest = "^7.2.0"
schedule = "^1.1.0"
apscheduler = "^3.9.1.post1"
black = "^22.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
